+++ native-frontend/db.js
@@ -5,3 +5,7 @@ export function save(key, value) {
 export function read(key) {
   return JSON.parse(window.localStorage.getItem(key));
 }
+
+export function clear(key) {
+  window.localStorage.setItem(key, null);
+}
